% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyComponent.R
\name{ShinyComponent}
\alias{ShinyComponent}
\title{Create A Shiny Component from an R Markdown Document}
\description{
Create A Shiny Component from an R Markdown Document

Create A Shiny Component from an R Markdown Document
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShinyComponent$new()}}
\item \href{#method-ui}{\code{ShinyComponent$ui()}}
\item \href{#method-server}{\code{ShinyComponent$server()}}
\item \href{#method-assets}{\code{ShinyComponent$assets()}}
\item \href{#method-app}{\code{ShinyComponent$app()}}
\item \href{#method-clone}{\code{ShinyComponent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyComponent$new(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyComponent$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyComponent$server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assets"></a>}}
\if{latex}{\out{\hypertarget{method-assets}{}}}
\subsection{Method \code{assets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyComponent$assets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-app"></a>}}
\if{latex}{\out{\hypertarget{method-app}{}}}
\subsection{Method \code{app()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyComponent$app(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyComponent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
